Update AddAndConfigure code
