+++ Sources/StytchCore/Documentation.docc/Resources/Code/AddAndConfigure/02-add-and-configure-03.swift
@@ -1,0 +1,30 @@
+import UIKit
+
+@UIApplicationMain
+final class AppDelegate: NSObject, UIApplicationDelegate {
+    private let stytchPublicToken = "your-public-token" // You might fetch this from your backend
+
+    func application(
+        _ application: UIApplication,
+        didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey : Any]? = nil
+    ) -> Bool {
+        StytchClient.configure(publicToken: publicToken)
+        return true
+    }
+
+    func application(
+        _ application: UIApplication,
+        continue userActivity: NSUserActivity,
+        restorationHandler: @escaping ([UIUserActivityRestoring]?) -> Void
+    ) -> Bool {
+        return true
+    }
+
+    func application(
+        _ app: UIApplication,
+        open url: URL,
+        options: [UIApplication.OpenURLOptionsKey : Any] = [:]
+    ) -> Bool {
+        return true
+    }
+}
